Retouch binarization and modify hidden layers.
Right now, with the prints we can see that metrics and system works as it should. Simply binarization is too high for too little learning.
Its basically predicting all under 0.5, so they would all go to 0. Since there's around 75% of 0 (NO), and that's more or less
the accuracy it's giving. The problem seems to be that it's not learning enough.